Add minimum expiry duration for timeseries
It can be set on the command line as a flag and will act as a minimum time
to wait after the last received data for a timeseries before expiring that
timeseries from the job runner.

This is useful if you have fast jobs where data might sometimes get sporatic
or delayed and you don't want the timeseries to expire and cause an error in the
K8s HPA system.